// src/app/model/teacher.ts
export class Teacher {

        id: number;
        ime: string;
        prezime: string;
        adresa: string;
        telefon: string;
        datumRodjenja: string;
        pol: string;
        username: string;
        password: string;
        mestoId: number;
        nastavnikTipId: number;

}